modified:   public/1.php

+++ public/1.php
@@ -1,6 +1,6 @@
 <?php
 $baht_thai    = "42.41";
 $order_price  = "4157";
-$real_amt_bath = $order_price / $baht_thai; 
+$real_amt_bath = (int)($order_price / $baht_thai); 
 
 echo $real_amt_bath;
